requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
httpx==0.24.1
anthropic==0.8.1
supabase==2.3.0
pydantic==2.5.3
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.111.0

# Security dependencies
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
slowapi==0.1.9

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.24.1

# Development dependencies
black==23.12.1
ruff==0.1.9
mypy==1.8.0